Split discount tests
