+++ saleor/graphql/discount/mutations/bulk_mutations.py
@@ -1,13 +1,13 @@
 import graphene
 
-from ...core.permissions import DiscountPermissions
-from ...discount import models
-from ...discount.utils import fetch_catalogue_info
-from ..core.mutations import ModelBulkDeleteMutation
-from ..core.types import DiscountError, NonNullList
-from ..plugins.dataloaders import load_plugin_manager
-from .mutations.utils import convert_catalogue_info_to_global_ids
-from .types import Sale, Voucher
+from ....core.permissions import DiscountPermissions
+from ....discount import models
+from ....discount.utils import fetch_catalogue_info
+from ...core.mutations import ModelBulkDeleteMutation
+from ...core.types import DiscountError, NonNullList
+from ...plugins.dataloaders import load_plugin_manager
+from ..types import Sale, Voucher
+from .utils import convert_catalogue_info_to_global_ids
 
 
 class SaleBulkDelete(ModelBulkDeleteMutation):
