+++ saleor/graphql/discount/tests/mutations/test_sale_bulk_delete.py
@@ -4,11 +4,11 @@ from unittest import mock
 import graphene
 import pytest
 
-from ....discount.models import Sale, SaleChannelListing, Voucher, VoucherChannelListing
-from ....discount.utils import fetch_catalogue_info
-from ....webhook.payloads import generate_sale_payload
-from ...tests.utils import get_graphql_content
-from ..mutations.utils import convert_catalogue_info_to_global_ids
+from .....discount.models import Sale, SaleChannelListing
+from .....discount.utils import fetch_catalogue_info
+from .....webhook.payloads import generate_sale_payload
+from ....tests.utils import get_graphql_content
+from ...mutations.utils import convert_catalogue_info_to_global_ids
 
 
 @pytest.fixture
@@ -23,40 +23,6 @@ def sale_list(channel_USD):
         ]
     )
     return list(sales)
-
-
-@pytest.fixture
-def voucher_list(channel_USD):
-    [voucher_1, voucher_2, voucher_3] = Voucher.objects.bulk_create(
-        [
-            Voucher(code="voucher-1"),
-            Voucher(code="voucher-2"),
-            Voucher(code="voucher-3"),
-        ]
-    )
-    VoucherChannelListing.objects.bulk_create(
-        [
-            VoucherChannelListing(
-                voucher=voucher_1,
-                channel=channel_USD,
-                discount_value=1,
-                currency=channel_USD.currency_code,
-            ),
-            VoucherChannelListing(
-                voucher=voucher_2,
-                channel=channel_USD,
-                discount_value=2,
-                currency=channel_USD.currency_code,
-            ),
-            VoucherChannelListing(
-                voucher=voucher_3,
-                channel=channel_USD,
-                discount_value=3,
-                currency=channel_USD.currency_code,
-            ),
-        ]
-    )
-    return voucher_1, voucher_2, voucher_3
 
 
 SALE_BULK_DELETE_MUTATION = """
@@ -166,64 +132,3 @@ def test_delete_sales_with_variants_triggers_webhook(
 
     assert content["data"]["saleBulkDelete"]["count"] == 3
     assert mocked_webhook_trigger.call_count == 3
-
-
-BULK_DELETE_VOUCHERS_MUTATION = """
-    mutation voucherBulkDelete($ids: [ID!]!) {
-        voucherBulkDelete(ids: $ids) {
-            count
-        }
-    }
-"""
-
-
-def test_delete_vouchers(staff_api_client, voucher_list, permission_manage_discounts):
-    variables = {
-        "ids": [
-            graphene.Node.to_global_id("Voucher", voucher.id)
-            for voucher in voucher_list
-        ]
-    }
-    response = staff_api_client.post_graphql(
-        BULK_DELETE_VOUCHERS_MUTATION,
-        variables,
-        permissions=[permission_manage_discounts],
-    )
-    content = get_graphql_content(response)
-
-    assert content["data"]["voucherBulkDelete"]["count"] == 3
-    assert not Voucher.objects.filter(
-        id__in=[voucher.id for voucher in voucher_list]
-    ).exists()
-
-
-@mock.patch("saleor.plugins.webhook.plugin.get_webhooks_for_event")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
-def test_delete_vouchers_trigger_webhook(
-    mocked_webhook_trigger,
-    mocked_get_webhooks_for_event,
-    any_webhook,
-    staff_api_client,
-    voucher_list,
-    permission_manage_discounts,
-    settings,
-):
-    # given
-    mocked_get_webhooks_for_event.return_value = [any_webhook]
-    settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
-
-    variables = {
-        "ids": [
-            graphene.Node.to_global_id("Voucher", voucher.id)
-            for voucher in voucher_list
-        ]
-    }
-    response = staff_api_client.post_graphql(
-        BULK_DELETE_VOUCHERS_MUTATION,
-        variables,
-        permissions=[permission_manage_discounts],
-    )
-    content = get_graphql_content(response)
-
-    assert content["data"]["voucherBulkDelete"]["count"] == 3
-    assert mocked_webhook_trigger.call_count == len(voucher_list)
