+++ saleor/graphql/discount/tests/mutations/test_voucher_channel_listing_update.py
@@ -5,12 +5,12 @@ import graphene
 from django.utils.functional import SimpleLazyObject
 from freezegun import freeze_time
 
-from ....core.utils.json_serializer import CustomJsonEncoder
-from ....discount import DiscountValueType
-from ....discount.error_codes import DiscountErrorCode
-from ....webhook.event_types import WebhookEventAsyncType
-from ....webhook.payloads import generate_meta, generate_requestor
-from ...tests.utils import assert_no_permission, get_graphql_content
+from .....core.utils.json_serializer import CustomJsonEncoder
+from .....discount import DiscountValueType
+from .....discount.error_codes import DiscountErrorCode
+from .....webhook.event_types import WebhookEventAsyncType
+from .....webhook.payloads import generate_meta, generate_requestor
+from ....tests.utils import assert_no_permission, get_graphql_content
 
 VOUCHER_CHANNEL_LISTING_UPDATE_MUTATION = """
 mutation UpdateVoucherChannelListing($id: ID!, $input: VoucherChannelListingInput!) {
