+++ saleor/graphql/discount/tests/mutations/test_voucher_delete.py
@@ -1,0 +1,90 @@
+import json
+from unittest.mock import patch
+
+import graphene
+import pytest
+from django.utils.functional import SimpleLazyObject
+from freezegun import freeze_time
+
+from .....core.utils.json_serializer import CustomJsonEncoder
+from .....webhook.event_types import WebhookEventAsyncType
+from .....webhook.payloads import generate_meta, generate_requestor
+from ....tests.utils import get_graphql_content
+
+VOUCHER_DELETE_MUTATION = """
+    mutation DeleteVoucher($id: ID!) {
+        voucherDelete(id: $id) {
+            voucher {
+                name
+                id
+            }
+            errors {
+                field
+                code
+                message
+            }
+          }
+        }
+"""
+
+
+def test_voucher_delete_mutation(
+    staff_api_client, voucher, permission_manage_discounts
+):
+    variables = {"id": graphene.Node.to_global_id("Voucher", voucher.id)}
+
+    response = staff_api_client.post_graphql(
+        VOUCHER_DELETE_MUTATION, variables, permissions=[permission_manage_discounts]
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["voucherDelete"]
+    assert data["voucher"]["name"] == voucher.name
+    with pytest.raises(voucher._meta.model.DoesNotExist):
+        voucher.refresh_from_db()
+
+
+@freeze_time("2022-05-12 12:00:00")
+@patch("saleor.plugins.webhook.plugin.get_webhooks_for_event")
+@patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
+def test_voucher_delete_mutation_trigger_webhook(
+    mocked_webhook_trigger,
+    mocked_get_webhooks_for_event,
+    any_webhook,
+    staff_api_client,
+    voucher,
+    permission_manage_discounts,
+    settings,
+):
+    # given
+    mocked_get_webhooks_for_event.return_value = [any_webhook]
+    settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
+
+    variables = {"id": graphene.Node.to_global_id("Voucher", voucher.id)}
+
+    # when
+    response = staff_api_client.post_graphql(
+        VOUCHER_DELETE_MUTATION, variables, permissions=[permission_manage_discounts]
+    )
+    content = get_graphql_content(response)
+
+    # then
+    assert content["data"]["voucherDelete"]["voucher"]
+    mocked_webhook_trigger.assert_called_once_with(
+        json.dumps(
+            {
+                "id": variables["id"],
+                "name": voucher.name,
+                "code": voucher.code,
+                "meta": generate_meta(
+                    requestor_data=generate_requestor(
+                        SimpleLazyObject(lambda: staff_api_client.user)
+                    )
+                ),
+            },
+            cls=CustomJsonEncoder,
+        ),
+        WebhookEventAsyncType.VOUCHER_DELETED,
+        [any_webhook],
+        voucher,
+        SimpleLazyObject(lambda: staff_api_client.user),
+    )
