+++ saleor/graphql/discount/tests/queries/test_vouchers_sorting.py
@@ -1,9 +1,10 @@
 import graphene
 import pytest
+from django.utils import timezone
 
-from ....discount import DiscountValueType, VoucherType
-from ....discount.models import Voucher, VoucherChannelListing
-from ...tests.utils import assert_graphql_error_with_message, get_graphql_content
+from .....discount import DiscountValueType, VoucherType
+from .....discount.models import Voucher, VoucherChannelListing
+from ....tests.utils import assert_graphql_error_with_message, get_graphql_content
 
 
 @pytest.fixture
@@ -344,3 +345,102 @@ def test_vouchers_with_filter_by_ids_and_channel_USD(
     assert len(vouchers_nodes) == len(vouchers)
     for voucher in vouchers_nodes:
         assert voucher["node"]["id"] in ids
+
+
+QUERY_VOUCHER_WITH_SORT = """
+    query ($sort_by: VoucherSortingInput!) {
+        vouchers(first:5, sortBy: $sort_by) {
+            edges{
+                node{
+                    name
+                }
+            }
+        }
+    }
+"""
+
+
+@pytest.mark.parametrize(
+    "voucher_sort, result_order",
+    [
+        (
+            {"field": "CODE", "direction": "ASC"},
+            ["Voucher2", "Voucher1", "FreeShipping"],
+        ),
+        (
+            {"field": "CODE", "direction": "DESC"},
+            ["FreeShipping", "Voucher1", "Voucher2"],
+        ),
+        (
+            {"field": "TYPE", "direction": "ASC"},
+            ["Voucher1", "Voucher2", "FreeShipping"],
+        ),
+        (
+            {"field": "TYPE", "direction": "DESC"},
+            ["FreeShipping", "Voucher2", "Voucher1"],
+        ),
+        (
+            {"field": "START_DATE", "direction": "ASC"},
+            ["FreeShipping", "Voucher2", "Voucher1"],
+        ),
+        (
+            {"field": "START_DATE", "direction": "DESC"},
+            ["Voucher1", "Voucher2", "FreeShipping"],
+        ),
+        (
+            {"field": "END_DATE", "direction": "ASC"},
+            ["Voucher2", "FreeShipping", "Voucher1"],
+        ),
+        (
+            {"field": "END_DATE", "direction": "DESC"},
+            ["Voucher1", "FreeShipping", "Voucher2"],
+        ),
+        (
+            {"field": "USAGE_LIMIT", "direction": "ASC"},
+            ["Voucher1", "FreeShipping", "Voucher2"],
+        ),
+        (
+            {"field": "USAGE_LIMIT", "direction": "DESC"},
+            ["Voucher2", "FreeShipping", "Voucher1"],
+        ),
+    ],
+)
+def test_query_vouchers_with_sort(
+    voucher_sort, result_order, staff_api_client, permission_manage_discounts
+):
+    Voucher.objects.bulk_create(
+        [
+            Voucher(
+                name="Voucher1",
+                code="abc",
+                discount_value_type=DiscountValueType.FIXED,
+                type=VoucherType.ENTIRE_ORDER,
+                usage_limit=10,
+            ),
+            Voucher(
+                name="Voucher2",
+                code="123",
+                discount_value_type=DiscountValueType.FIXED,
+                type=VoucherType.ENTIRE_ORDER,
+                start_date=timezone.now().replace(year=2012, month=1, day=5),
+                end_date=timezone.now().replace(year=2013, month=1, day=5),
+            ),
+            Voucher(
+                name="FreeShipping",
+                code="xyz",
+                discount_value_type=DiscountValueType.PERCENTAGE,
+                type=VoucherType.SHIPPING,
+                start_date=timezone.now().replace(year=2011, month=1, day=5),
+                end_date=timezone.now().replace(year=2015, month=12, day=31),
+                usage_limit=1000,
+            ),
+        ]
+    )
+    variables = {"sort_by": voucher_sort}
+    staff_api_client.user.user_permissions.add(permission_manage_discounts)
+    response = staff_api_client.post_graphql(QUERY_VOUCHER_WITH_SORT, variables)
+    content = get_graphql_content(response)
+    vouchers = content["data"]["vouchers"]["edges"]
+
+    for order, voucher_name in enumerate(result_order):
+        assert vouchers[order]["node"]["name"] == voucher_name
